Add files via upload
This code is the Windows side of the DCC controller refered to as KATS. This is the first working version. As of this release, the comm port and the locomotive address can be seleceted which are used to control the speed and direction of a locomotive thru the DCC station. The DCC station is a STM32 based controller